Stream task15 grouping [1 of 20]

+++ src/stream/task14_summarystatistics_10/Main20.java
@@ -50,7 +50,7 @@ public class Main20 {
         return Arrays.stream(words).collect(Collectors.groupingBy(w->w.charAt(0), Collectors.averagingInt(String::length)));
     }
 
-    public Map <Character, Integer> practice7 () {
+    public Map<Character, Long> practice7 () {
         String[] words = {"apple", "orange",
                 "banana", "umbrella"};
         return Arrays.stream(words).flatMap(w->w.chars().mapToObj(c->(char)c)).filter(c->"aeiou".indexOf(c)>=0).collect(Collectors.groupingBy(Function.identity(), Collectors.counting()));
